[Fix] Terminated sessions appear in volunteer session list (WIP) (#93)
* added logic to display message when user joins and leaves whiteboard session.

* Add autocorrect code.
Also prevent throwing errors if the number already in the
database is not in the right format.

* correct name for parameter

* Don't print out the whole user model!

* Added some clarifying comments, and also moved regex explanation

* Check if phonePretty is being set to null or undefined.

* spelling

* Changed pruneDeadSessions definition syntax

The bug happened because of the difference in behavior of the
this keyword when a function is defined by arrow vs. anonymous
function syntax. With the arrow syntax, the check for
this._sessions always failed and the dead sessions were never pruned.

* Changed logic for socket association.

People who logged out but still have the socket should stay connected.

* minor style fix

* Changed logic for disconnecting old sockets

This way we ensure that
SessionManager.disconnect() doesn't
disconnect users who have new socket objects
only because they lost their connection.

* Removed joined/left messages